Refactor memory benchmark code
